pyproject.toml
[project]
name = "hyper-sentinel-v2"
version = "0.2.0"
description = "Next-generation autonomous fintech AI agent swarm"
requires-python = ">=3.12"
dependencies = [
    # Core agent frameworks
    "upsonic>=0.50.0",
    "agno>=1.0.0",
    # LLM providers
    "anthropic>=0.40.0",
    "openai>=1.40.0",
    "google-generativeai>=0.8.0",
    # Messaging
    "nats-py>=2.7.0",
    # Data models
    "pydantic>=2.5.0",
    # Environment
    "python-dotenv>=1.0.0",
    # Market data
    "yfinance>=0.2.30",
    "requests>=2.32.0",
    # Trading SDKs
    "hyperliquid-python-sdk>=0.22.0",
    "py-clob-client>=0.34.6",
    # News & intelligence
    "y2-py>=0.4.1",
    # Macro data
    "fredapi>=0.5.2",
    # Webhooks
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    # MCP
    "mcp>=1.0.0",
    # UI
    "rich>=13.0.0",
    # Retry + data
    "tenacity>=8.2.0",
    "pandas>=2.2.0",
    "websockets>=12.0",
    # Scheduler
    "schedule>=1.2.0",
    # Telegram notifications
    "python-telegram-bot>=21.0",
    # Technical analysis
    "pandas-ta>=0.3.14b",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
executionEnvironments = [{ root = "." }]
